9 Jan 10am

+++ src/day04/Calculator.java
@@ -2,9 +2,10 @@ package day04;
 
 public class Calculator {
     
-    public static double calc(int first, int second, String operand){
+    public static double calc(double first, double second, String operand){
 
         double output=0d;
+        
 
         if(operand.equals("+")){
             output = first + second;
